+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Ribbon V2.14.1, RibbonTools V1.6.0
+
+#### Changed (Ribbon)
+
+- Destroy EventLogger.
+
 ### Ribbon V2.14.0, RibbonTools V1.6.0
 
 #### Changed (Ribbon)